+++ packages/terra-table/CHANGELOG.md
@@ -3,8 +3,7 @@ Changelog
 
 Unreleased
 ----------
-### Changed
-* Changed background color to transparent
+
 ### Added
 * Added documentation pages for TableRows, SelectableTableRows, and SingleSelectableRows