+++ packages/terra-table/CHANGELOG.md
@@ -3,11 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Changed background color to transparent
+
 ### Added
 * Added documentation pages for TableRows, SelectableTableRows, and SingleSelectableRows
-
-### Changed
-* Updated documentation
 
 2.19.0 - (August 1, 2018)
 ------------------