+++ packages/terra-table/CHANGELOG.md
@@ -3,11 +3,14 @@ Changelog
 
 Unreleased
 ----------
-### Changed
-* Changed background color to transparent
 
+2.20.0 - (August 8, 2018)
+------------------
 ### Added
 * Added documentation pages for TableRows, SelectableTableRows, and SingleSelectableRows
+
+### Changed
+* Changed background color to transparent
 
 2.19.0 - (August 1, 2018)
 ------------------